% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lga_data}
\alias{lga_data}
\title{The 2022 population projection of LGAs in Abia state}
\format{
A data frame with 17 rows and 5 variables:
\describe{
  \item{state}{State of data}
  \item{lga}{Local Government Areas of the listed state}
  \item{population}{the 2022 population estimate per LGA}
  \item{prev_x}{random prevalence of a condition "x" created by the authors}
   \item{incidence_x}{random annual incidence of a condition "x" created by the authors}
  }
}
\usage{
lga_data
}
\description{
A minimal 2022 population estimate of LGAs in Abia state, Nigeria, as obtained from the
[City Population website]("https://www.citypopulation.de/en/nigeria/admin/"). The data has been pre-cleaned and two
 additional columns `prev_x` and `incidence_x` added based on random calculations.
}
\note{
for more information, kindly visit \url{https://www.citypopulation.de/en/nigeria/admin/}
}
\keyword{datasets}
