% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_lgas2.R
\name{map_lgas2}
\alias{map_lgas2}
\title{Plot LGA-level Map with Bubbles}
\usage{
map_lgas2(
  .data,
  fill,
  bubble,
  st = .data$state,
  lga = lga,
  label = FALSE,
  cols = NULL,
  size = NULL,
  interactive = FALSE
)
}
\arguments{
\item{.data}{a tabular (rectangular) data containing LGA and the variables to be plotted}

\item{fill}{the variable (categorical or continuous) to be used to fill (color) the LGAs}

\item{bubble}{the variable name of the column to be used for the bubble (point) plot}

\item{st}{the variable name of the column representing states in the data}

\item{lga}{the variable name of the column representing LGAs in the data}

\item{label}{logical (boolean), indicating if the LGAs should be labelled or not}

\item{cols}{a string of colors equal to the number of categories contained in the `fill` variable. When this is supplied, it overides the default colors and allow the user to choose specific colors}

\item{size}{the text size of the labels in whole numbers. The default is 2}

\item{interactive}{logical (boolean), indicating whether the map should allow some level of interactivity}
}
\value{
a two-dimensional LGA-level map
}
\description{
Plot LGA-level Map with Bubbles
}
\examples{
NULL
}
