% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_states3.R
\name{map_states3}
\alias{map_states3}
\title{Plot State-level Map with Two Bubbles}
\usage{
map_states3(
  .data,
  fill,
  light_b,
  dark_b,
  state = state,
  label = FALSE,
  cols = NULL,
  size = NULL,
  interactive = FALSE
)
}
\arguments{
\item{.data}{a tabular (rectangular) data containing LGA and the variables to be plotted}

\item{fill}{the variable (categorical or continuous) to be used to fill (color) the LGAs}

\item{light_b}{the variable name of the column to be used for the lighter (bigger) bubble}

\item{dark_b}{the variable name of the column to be used for the darker (smaller) bubble}

\item{state}{the variable name of the column representing states in the data}

\item{label}{logical (boolean), indicating if the LGAs should be labelled or not}

\item{cols}{a string of colors equal to the number of categories contained in the `fill` variable. When this is supplied, it overides the default colors and allow the user to choose specific colors}

\item{size}{the text size of the labels in whole numbers. The default is 2}

\item{interactive}{logical (boolean), indicating whether the map should allow some level of interactivity}
}
\value{
a three-dimensional state-level map
}
\description{
Plot State-level Map with Two Bubbles
}
\examples{

## map 2022 spectrum estimate for "Ondo", "Oyo", "Osun" and "Ogun" states interactively

state_data <- spectrum(state = c("Ondo", "Oyo", "Osun", "Ogun")) |>
  dplyr::count(state, wt = estimate, name = "estimate") |>
  dplyr::mutate(estimate_scaled = estimate * 0.8)

map_states3(state_data, fill = estimate, light_b = estimate, dark_b = estimate_scaled, label = TRUE)

}
