% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\name{my_cols}
\alias{my_cols}
\title{Selected Colors for Charts}
\usage{
my_cols(n)
}
\arguments{
\item{n}{number of colors. This should be equal to the number of categories contained in the data}
}
\value{
selected colors
}
\description{
Selected Colors for Charts
}
\examples{
NULL
}
\keyword{internal}
