% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr_lgas.R
\name{ndr_lgas}
\alias{ndr_lgas}
\title{Coordinates of Nigeria LGAs}
\usage{
ndr_lgas(state)
}
\arguments{
\item{state}{string (character). The names of state (or LGA)}
}
\value{
coordinates of selected LGAs
}
\description{
Coordinates of Nigeria LGAs
}
\examples{

## get LGA coordinates for Gombe state

gom_lga_coord <- ndr_lgas("Gombe")

## get LGA coordinates for Anambra and Kogi
ana_kog_coord <- ndr_lgas(c("Anambra", "Kogi"))

}
