% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr_states.R
\name{ndr_states}
\alias{ndr_states}
\title{Coordinates of Nigeria States}
\usage{
ndr_states(state)
}
\arguments{
\item{state}{string (character). The names of state (or LGA)}
}
\value{
coordinates of selected states
}
\description{
Coordinates of Nigeria States
}
\examples{
NULL
}
