% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_lgas3.R
\name{map_lgas3}
\alias{map_lgas3}
\title{Plot LGA-level Map with Two Bubbles}
\usage{
map_lgas3(
  .data,
  fill,
  light_b,
  dark_b,
  state = state,
  lga = lga,
  label = FALSE,
  cols = NULL,
  size = NULL,
  interactive = FALSE
)
}
\arguments{
\item{.data}{a tabular (rectangular) data containing LGA and the variables to be plotted}

\item{fill}{the variable (categorical or continuous) to be used to fill (color) the LGAs}

\item{light_b}{the variable name of the column to be used for the lighter (bigger) bubble}

\item{dark_b}{the variable name of the column to be used for the darker (smaller) bubble}

\item{state}{the variable name of the column representing states in the data}

\item{lga}{the variable name of the column representing LGAs in the data}

\item{label}{logical (boolean), indicating if the LGAs should be labelled or not}

\item{cols}{a string of colors equal to the number of categories contained in the `fill` variable. When this is supplied, it overides the default colors and allow the user to choose specific colors}

\item{size}{the text size of the labels in whole numbers. The default is 2}

\item{interactive}{logical (boolean), indicating whether the map should allow some level of interactivity}
}
\value{
a three-dimensional LGA-level map
}
\description{
Plot LGA-level Map with Two Bubbles
}
\examples{

## map the lga_data, filling by `prev_x` and light_b by `population` and dark_b by `incidence_x`

map_lgas3(lga_data, fill = prev_x, light_b = population, dark_b = incidence_x, label = TRUE)

}
