% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_lgas3.R
\name{map_lgas3}
\alias{map_lgas3}
\title{Plot LGA-level Map with Two Bubbles}
\usage{
map_lgas3(
  .data,
  fill,
  big_bubble,
  small_bubble,
  state = state,
  lga = lga,
  label_lga = TRUE,
  label_fill = FALSE,
  label_small_bubble = TRUE,
  label_lga_color = NULL,
  label_fill_color = NULL,
  label_small_bubble_color = NULL,
  fill_colors = NULL,
  big_bubble_color = NULL,
  small_bubble_color = NULL,
  bubble_transparency = NULL,
  size_lga = NULL,
  size_fill = NULL,
  size_small_bubble = NULL,
  border_color = NULL,
  border_width = NULL,
  gradient = NULL,
  grad_direction = NULL,
  na_fill = NULL,
  all_regions = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{.data}{a tabular (rectangular) data containing LGA and the variables to be plotted}

\item{fill}{the variable (categorical or continuous) to be used to fill (color) the LGAs}

\item{big_bubble}{the variable name of the column to be used for the lighter (bigger) bubble}

\item{small_bubble}{the variable name of the column to be used for the darker (smaller) bubble}

\item{state}{the variable name of the column representing states in the data}

\item{lga}{the variable name of the column representing LGAs in the data}

\item{label_lga}{logical (boolean), indicating if the LGAs should be labelled or not}

\item{label_fill}{logical (boolean), indicating if the values for the `fill` data should be displayed}

\item{label_small_bubble}{logical (boolean), indicating if the values for the `small_bubble` data should be displayed}

\item{label_lga_color}{when supplied, this color will be used as font-color for the `lga` label}

\item{label_fill_color}{when supplied, this color will be used as font-color for the `fill` label}

\item{label_small_bubble_color}{when supplied, this color will be used as font-color for the `small_bubble` label}

\item{fill_colors}{a string of colors equal to the number of categories contained in the `fill` variable. When this is supplied, it overides the default colors and allow the user to choose specific colors}

\item{big_bubble_color}{a string of color. When this is supplied, it overides the default color for the `big_bubble` plots}

\item{small_bubble_color}{a string of color. When this is supplied, it overides the default color for the `small_bubble` plots}

\item{bubble_transparency}{the alpha transparency of the bubbles. Value ranges from 0 (transparent) to 1(solid)}

\item{size_lga}{the font-size for the `lga` labels in whole numbers.}

\item{size_fill}{the font-size for the `fill` labels in whole numbers.}

\item{size_small_bubble}{the font-size for the `small_bubble` labels in whole numbers. The default is 2}

\item{border_color}{the color to be used for the map boundaries. The default is black}

\item{border_width}{the thickness of the border line numbers}

\item{gradient}{when map is filled by a continuous variable, a gradient color can be supplied. The options are "A" to "H". The default is "E"}

\item{grad_direction}{the gradient direction can be reversed by negating the current value. The default is `-1`. The reverse will take a value of `1`}

\item{na_fill}{the fill color to be used for locations with missing value. The default is `pink`}

\item{all_regions}{logical (boolean), indicating if the Nigeria LGAs should be used for the map}

\item{interactive}{logical (boolean), indicating whether the map should allow some level of interactivity}
}
\value{
a three-dimensional LGA-level map
}
\description{
Plot LGA-level Map with Two Bubbles
}
\examples{

## map the lga_data, filling by `prev_x` and light_b by `population` and dark_b by `incidence_x`

map_lgas3(
  lga_data,
  fill = prev_x,
  big_bubble = population,
  small_bubble = incidence_x,
  label_small_bubble = TRUE
)

}
