% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_states.R
\name{map_states}
\alias{map_states}
\title{Plot State-level Map}
\usage{
map_states(
  .data,
  fill,
  st = state,
  label = FALSE,
  cols = NULL,
  size = NULL,
  interactive = FALSE
)
}
\arguments{
\item{.data}{a tabular (rectangular) data containing LGA and the variables to be plotted}

\item{fill}{the variable (categorical or continuous) to be used to fill (color) the LGAs}

\item{st}{the variable name of the column representing states in the data}

\item{label}{logical (boolean), indicating if the LGAs should be labelled or not}

\item{cols}{a string of colors equal to the number of categories contained in the `fill` variable. When this is supplied, it overides the default colors and allow the user to choose specific colors}

\item{size}{the text size of the labels in whole numbers. The default is 2}

\item{interactive}{logical (boolean), indicating whether the map should allow some level of interactivity}
}
\value{
state-level map
}
\description{
Plot State-level Map
}
\examples{
NULL
}
