% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_states3.R
\name{map_states3}
\alias{map_states3}
\title{Plot State-level Map with Two Bubbles}
\usage{
map_states3(
  .data,
  fill,
  light_bubble,
  dark_bubble,
  state = state,
  label_state = TRUE,
  label_fill = FALSE,
  label_dark_bubble = FALSE,
  label_state_color = NULL,
  label_fill_color = NULL,
  label_dark_bubble_color = NULL,
  fill_colors = NULL,
  light_bubble_color = NULL,
  dark_bubble_color = NULL,
  size_state = NULL,
  size_fill = NULL,
  size_dark_bubble = NULL,
  border_color = NULL,
  border_width = NULL,
  gradient = NULL,
  grad_direction = NULL,
  na_fill = NULL,
  interactive = FALSE
)
}
\arguments{
\item{.data}{a tabular (rectangular) data containing LGA and the variables to be plotted}

\item{fill}{the variable (categorical or continuous) to be used to fill (color) the LGAs}

\item{light_bubble}{the variable name of the column to be used for the lighter (bigger) bubble}

\item{dark_bubble}{the variable name of the column to be used for the darker (smaller) bubble}

\item{state}{the variable name of the column representing states in the data}

\item{label_state}{logical (boolean), indicating if the state should be labelled or not}

\item{label_fill}{logical (boolean), indicating if the values for the `fill` data should be displayed}

\item{label_dark_bubble}{logical (boolean), indicating if the values for the `dark_bubble` data should be displayed}

\item{label_state_color}{when supplied, this color will be used as font-color for the `state` label}

\item{label_fill_color}{when supplied, this color will be used as font-color for the `fill` label}

\item{label_dark_bubble_color}{when supplied, this color will be used as font-color for the `dark_bubble` label}

\item{fill_colors}{a string of colors equal to the number of categories contained in the `fill` variable. When this is supplied, it overides the default colors and allow the user to choose specific colors}

\item{light_bubble_color}{a string of color. When this is supplied, it overides the default color for the `light_bubble` plots}

\item{dark_bubble_color}{a string of color. When this is supplied, it overides the default color for the `dark_bubble` plots}

\item{size_state}{the font-size for the `state` labels in whole numbers.}

\item{size_fill}{the font-size for the `fill` labels in whole numbers.}

\item{size_dark_bubble}{the font-size for the `dark_bubble` labels in whole numbers. The default is 2}

\item{border_color}{the color to be used for the map boundaries. The default is black}

\item{border_width}{the thickness of the border line numbers}

\item{gradient}{when map is filled by a continuous variable, a gradient color can be supplied. The options are "A" to "H". The default is "E"}

\item{grad_direction}{the gradient direction can be reversed by negating the current value. The default is `-1`. The reverse will take a value of `1`}

\item{na_fill}{the fill color to be used for locations with missing value. The default is `pink`}

\item{interactive}{logical (boolean), indicating whether the map should allow some level of interactivity}
}
\value{
a three-dimensional state-level map
}
\description{
Plot State-level Map with Two Bubbles
}
\examples{

## map 2022 spectrum estimate for "Ondo", "Oyo", "Osun" and "Ogun" states interactively

state_data <- spectrum(state = c("Ondo", "Oyo", "Osun", "Ogun")) |>
  dplyr::count(state, wt = estimate, name = "estimate") |>
  dplyr::mutate(estimate_scaled = estimate * 0.8)

map_states3(state_data, fill = estimate, light_bubble = estimate, dark_bubble = estimate_scaled)

}
