% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapndr-package.R
\docType{package}
\name{mapndr-package}
\alias{mapndr}
\alias{mapndr-package}
\title{mapndr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
