% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr_lgas.R
\name{ndr_lgas}
\alias{ndr_lgas}
\title{Coordinates of Nigeria LGAs}
\usage{
ndr_lgas(state)
}
\arguments{
\item{state}{string (character). The names of state (or LGA)}
}
\value{
coordinates of selected LGAs
}
\description{
Coordinates of Nigeria LGAs
}
\examples{
NULL
}
