% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr_states.R
\name{ndr_states}
\alias{ndr_states}
\title{Coordinates of Nigeria States}
\usage{
ndr_states(state)
}
\arguments{
\item{state}{string (character). The names of state (or LGA)}
}
\value{
coordinates of selected states
}
\description{
Coordinates of Nigeria States
}
\examples{

## get the coordinates of Zamfara state

zam_coord <- ndr_states("Zamfara")

## get the coordinates of Bayelsa and Ebonyi states
bay_ebo_coords <- ndr_states(c("Bayelsa", "Ebonyi"))
}
