% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pyramid.R
\name{plot_pyramid}
\alias{plot_pyramid}
\title{Pyramid Plot of Age and Sex Data}
\usage{
plot_pyramid(
  .data,
  var,
  age_band = .data$age_group,
  sex = sex,
  label = TRUE,
  cols = NULL,
  size = NULL,
  interactive = FALSE
)
}
\arguments{
\item{.data}{a tabular (rectangular) data containing LGA and the variables to be plotted}

\item{var}{the variable name in the data to be used to plot the pyramid}

\item{age_band}{the variable name in the data that contains the different age-band}

\item{sex}{the variable name that contains the sex (gender) in the supplied data. The `sex` should be in the "M" and "F" format as commonly seen on the NDR}

\item{label}{logical (boolean), indicating if the LGAs should be labelled or not}

\item{cols}{a string of colors equal to the number of categories contained in the `fill` variable. When this is supplied, it overides the default colors and allow the user to choose specific colors}

\item{size}{the text size of the labels in whole numbers. The default is 2}

\item{interactive}{logical (boolean), indicating whether the map should allow some level of interactivity}
}
\value{
population pyramid plot
}
\description{
Pyramid Plot of Age and Sex Data
}
\examples{
NULL
}
