% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pyramid2.R
\name{plot_pyramid2}
\alias{plot_pyramid2}
\title{Pyramid Plot of Age and Sex Data}
\usage{
plot_pyramid2(
  .data,
  light_fill,
  dark_fill,
  age_group = age_group,
  sex = sex,
  digits = NULL,
  age_bands = NULL,
  label = TRUE,
  fill_colors = NULL,
  size = NULL,
  border = TRUE,
  border_color = NULL,
  inverse = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{.data}{a tabular (rectangular) data containing LGA and the variables to be plotted}

\item{light_fill}{the variable name in the data to be used to fill the bar (the lighter color)}

\item{dark_fill}{the variable name of the smaller values for the darker (shorter) fill}

\item{age_group}{the variable name in the data that contains the different age-band}

\item{sex}{the variable name that contains the sex (gender) in the supplied data. The `sex` should be in the "M" and "F" format as commonly seen on the NDR}

\item{digits}{The number of significant numbers to be used for the proportion/ratio}

\item{age_bands}{a vector of age-bands contained in the data to be used for age-categorization, if different from
c( "0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59", "60-64", "65+")}

\item{label}{logical (boolean), indicating if the data should be labelled or not}

\item{fill_colors}{a string of colors equal to two, in the number of unique entries in the `sex` variable}

\item{size}{The font-size of the `fill` labels in whole numbers}

\item{border}{logical (boolean). It determines if the individual bars should have a border or not}

\item{border_color}{if the `border` argument is set to `TRUE`, the user can supplied a border color using any of the standard color formats}

\item{inverse}{logical (boolean). If set to `TRUE`, the pyramid plot will be inverted, starting from highest age-group to lowest}

\item{interactive}{logical (boolean), indicating whether the map should allow some level of interactivity}
}
\value{
population pyramid plot
}
\description{
Pyramid Plot of Age and Sex Data
}
\examples{

## plot 2022 spectrum estimate of Zamfara by sex and `incidence_x`

pop_data <- spectrum(year = 2022, state = "Zamfara") |>
  dplyr::mutate(
    age_group = forcats::fct_collapse(age_group, "65+" = c("65-69", "70-74", "75-79", "80+"))
  ) |>
  dplyr::group_by(state, sex, age_group) |>
  dplyr::summarise(
    estimate = sum(estimate, na.rm = TRUE),
    .groups = "drop"
  ) |>
  dplyr::mutate(
    incidence_x = round(estimate * runif(1, max = 0.3))
  )

plot_pyramid2(pop_data, light_fill = estimate, dark_fill = incidence_x)

}
