% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum}
\alias{spectrum}
\title{The Harvested Spectrum Data put together by the Authors}
\usage{
spectrum(major = 6, minor = 19, year = 2022, state = NULL)
}
\arguments{
\item{major}{optional. The major version number of the spectrum. Currently only contains version 6}

\item{minor}{optional. The minor version number of the spectrum. Currently only contains version 19}

\item{year}{optional. The spectrum year to be returned}

\item{state}{optional. The state of interest.}
}
\value{
rectangular data of interest
}
\description{
The Harvested Spectrum Data put together by the Authors
}
\examples{
## get Abia state spectrum estimate for the different age-bands and gender in 2022
abia_2022 <- spectrum(year = 2022, state = "Abia")

## get spectrum estimate for Adamawa and Rivers in 2023
ada_riv_2023 <- spectrum(year = 2023, state = c("Adamawa", "Rivers"))
}
